+++ test/test_app_send.cpp
@@ -9,7 +9,6 @@ Author: 李柯鹏 <likepeng0418@163.com>
 #include <vector>
 #include <thread>
 #include <chrono>
-#include <glog/logging.h>
 #include "myframe/common.h"
 #include "myframe/log.h"
 #include "myframe/msg.h"
@@ -17,7 +16,7 @@ Author: 李柯鹏 <likepeng0418@163.com>
 #include "myframe/mod_manager.h"
 #include "myframe/app.h"
 
-#include "performance_test_config.h"
+#include "test_config.h"
 
 class EchoActorTest : public myframe::Actor {
  public:
@@ -27,22 +26,14 @@ class EchoActorTest : public myframe::Actor {
   }
 
   void Proc(const std::shared_ptr<const myframe::Msg>& msg) override {
-    if (msg->GetData() == "hello") {
-      auto re = std::make_shared<myframe::Msg>(
-        "resp:" + std::to_string(seq_++));
-      auto mailbox = GetMailbox();
-      mailbox->Send(msg->GetSrc(), std::move(re));
-    }
+    LOG(INFO) << "recv " << msg->GetSrc() << ":" << msg->GetData();
   }
-
- private:
-  int seq_{0};
 };
 
 int main() {
   auto log_dir =
       myframe::Common::GetAbsolutePath(MYFRAME_LOG_DIR).string();
-  myframe::InitLog(log_dir, "app_send_req_test");
+  myframe::InitLog(log_dir, "test_app_send");
 
   auto app = std::make_shared<myframe::App>();
   myframe::Arguments args;
@@ -61,7 +52,7 @@ int main() {
   auto actor = mod->CreateActorInst("class", "EchoActorTest", "1");
   app->AddActor(actor);
 
-  // 压力测试SendRequest函数
+  // 测试Send函数
   std::mutex mtx;
   int th_cnt = 5;
   int exit_th_cnt = 0;
@@ -69,19 +60,13 @@ int main() {
   std::vector<std::thread> th_vec;
   for (int i = 0; i < th_cnt; ++i) {
     th_vec.push_back(std::thread([&, i](){
-      while (app->GetState() != myframe::App::State::kRunning) {
-        std::this_thread::sleep_for(
-          std::chrono::milliseconds(100));
-      }
       int cnt = send_cnt;
       while (cnt--) {
-        auto msg = std::make_shared<myframe::Msg>("hello");
+        auto msg = std::make_shared<myframe::Msg>(
+          "world " + std::to_string(i));
         msg->SetDst("actor.EchoActorTest.1");
-        auto resp = app->SendRequest(std::move(msg));
-        if (resp == nullptr) {
-          continue;
-        }
-        LOG(INFO) << "thread " << i << " resp: " << resp->GetData();
+        app->Send(std::move(msg));
+        std::this_thread::sleep_for(std::chrono::milliseconds(1));
       }
       std::lock_guard<std::mutex> g(mtx);
       LOG(INFO) << "user thread " << i << " exit";
