+++ test/test_performance_trans1_fullspeed.cpp
@@ -17,7 +17,7 @@ Author: 李柯鹏 <likepeng0418@163.com>
 #include "myframe/mod_manager.h"
 #include "myframe/app.h"
 
-#include "performance_test_config.h"
+#include "test_config.h"
 
 class FullSpeedTransTest : public myframe::Actor {
  public:
@@ -81,7 +81,7 @@ class FullSpeedTransTest : public myframe::Actor {
 int main() {
   auto log_dir =
       myframe::Common::GetAbsolutePath(MYFRAME_LOG_DIR).string();
-  myframe::InitLog(log_dir, "performance_trans1_fullspeed_test");
+  myframe::InitLog(log_dir, "test_performance_trans1_fullspeed");
 
   auto app = std::make_shared<myframe::App>();
   myframe::Arguments args;
